backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4
python-multipart==0.0.20
